+++ assets/js/content/debug.js
@@ -38,7 +38,7 @@ window.Debug = (function() {
 			"Page": true,
 			"Progress": true,
 
-			"Sound": false,
+			"Sound": true,
 			"StatsDisplay": false,
 			"Stats": false,
 			"Skin": true,