+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Version 0.2.18
 
 * Adds instructions for block list ("disabled domains")
-* Improves battle console usability
+* Improves BattleConsole usability, Dialogue system
 * Adds Disguises and new consumables
 
 ### Version 0.2.17