+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Adds instructions for block list ("disabled domains")
 * Improves BattleConsole usability, Dialogue system
 * Adds Disguises and new consumables
+* Adds new Tally voice
 
 ### Version 0.2.17